+++ src/main/resources/db/migration/V3__Create_question_table.sql
@@ -1,11 +1,11 @@
 CREATE TABLE question
 (
-    id INT AUTO_INCREMENT PRIMARY KEY,
+    id bigint AUTO_INCREMENT PRIMARY KEY,
     title VARCHAR(50),
     description TEXT,
     gmt_create BIGINT,
     gmt_modified BIGINT,
-    creator INT,
+    creator bigint,
     comment_count INT DEFAULT 0,
     view_count INT DEFAULT 0,
     like_count INT DEFAULT 0,
